+++ src/rust/src/types.rs
@@ -308,8 +308,6 @@ pub static SHA1: LazyPyImport =
     LazyPyImport::new("cryptography.hazmat.primitives.hashes", &["SHA1"]);
 pub static SHA256: LazyPyImport =
     LazyPyImport::new("cryptography.hazmat.primitives.hashes", &["SHA256"]);
-pub static SHA384: LazyPyImport =
-    LazyPyImport::new("cryptography.hazmat.primitives.hashes", &["SHA384"]);
 pub static SHA512: LazyPyImport =
     LazyPyImport::new("cryptography.hazmat.primitives.hashes", &["SHA512"]);
 
@@ -387,27 +385,6 @@ pub static ECDSA: LazyPyImport =
     LazyPyImport::new("cryptography.hazmat.primitives.asymmetric.ec", &["ECDSA"]);
 pub static ECDH: LazyPyImport =
     LazyPyImport::new("cryptography.hazmat.primitives.asymmetric.ec", &["ECDH"]);
-pub static SECP256R1: LazyPyImport = LazyPyImport::new(
-    "cryptography.hazmat.primitives.asymmetric.ec",
-    &["SECP256R1"],
-);
-pub static SECP384R1: LazyPyImport = LazyPyImport::new(
-    "cryptography.hazmat.primitives.asymmetric.ec",
-    &["SECP384R1"],
-);
-pub static SECP521R1: LazyPyImport = LazyPyImport::new(
-    "cryptography.hazmat.primitives.asymmetric.ec",
-    &["SECP521R1"],
-);
-
-pub static X25519_PUBLIC_KEY: LazyPyImport = LazyPyImport::new(
-    "cryptography.hazmat.primitives.asymmetric.x25519",
-    &["X25519PublicKey"],
-);
-pub static X25519_PRIVATE_KEY: LazyPyImport = LazyPyImport::new(
-    "cryptography.hazmat.primitives.asymmetric.x25519",
-    &["X25519PrivateKey"],
-);
 
 pub static ED25519_PRIVATE_KEY: LazyPyImport = LazyPyImport::new(
     "cryptography.hazmat.primitives.asymmetric.ed25519",